pyproject.toml
[project]
name = "seedvc"
version = "2.0.0"
description = "Seed-VC V2: Zero-shot voice conversion"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "scipy>=1.13.1",
    "librosa>=0.10.2",
    "huggingface-hub>=0.28.1",
    "einops>=0.8.0",
    "pydub>=0.25.1",
    "transformers>=4.46.3",
    "soundfile>=0.12.1",
    "numpy>=1.26.4,<2",
    "hydra-core>=1.3.2",
    "omegaconf",
    "pyyaml",
    "accelerate",
    "munch",
    "jiwer",
    "sounddevice",
    "modelscope",
    "python-dotenv",
]

[project.optional-dependencies]
codec = ["descript-audio-codec"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["seedvc", "hf_utils"]
packages = ["configs", "configs.v2", "modules", "modules.v2", "modules.astral_quantization", "modules.campplus", "modules.bigvgan", "modules.bigvgan.alias_free_activation", "modules.bigvgan.alias_free_activation.cuda"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]
